+++ src/main/java/se/isselab/HAnS/featureView/FeatureViewElement.java
@@ -71,9 +71,8 @@ public class FeatureViewElement implements StructureViewTreeElement, SortableTre
         return presentation != null ? presentation : new PresentationData();
     }
 
-    @NotNull
     @Override
-    public TreeElement[] getChildren() {
+    public TreeElement @NotNull [] getChildren() {
         List<FeatureModelFeature> properties = PsiTreeUtil.getChildrenOfTypeAsList(myElement, FeatureModelFeature.class);
         if (!properties.isEmpty()) {
             List<TreeElement> treeElements = new ArrayList<>();