+++ src/main/java/se/isselab/HAnS/featureView/FeatureViewModel.java
@@ -31,8 +31,7 @@ public class FeatureViewModel extends StructureViewModelBase implements
     }
 
 
-    @NotNull
-    public Sorter[] getSorters() {
+    public Sorter @NotNull [] getSorters() {
         return new Sorter[]{Sorter.ALPHA_SORTER};
     }