+++ src/main/java/se/isselab/HAnS/syntaxHighlighting/HansColorSettingsPage.java
@@ -76,19 +76,19 @@ public class HansColorSettingsPage implements ColorSettingsPage {
         map.put("EA_key", CodeAnnotationsSyntaxHighlighter.KEYWORD);
         return map;
     }
-    @NotNull
+
     @Override
-    public AttributesDescriptor[] getAttributeDescriptors() {
+    public AttributesDescriptor @NotNull [] getAttributeDescriptors() {
         return DESCRIPTORS;
     }
-    @NotNull
+
     @Override
-    public ColorDescriptor[] getColorDescriptors() {
+    public ColorDescriptor @NotNull [] getColorDescriptors() {
         return ColorDescriptor.EMPTY_ARRAY;
     }
-    @NotNull
+
     @Override
-    public String getDisplayName() {
+    public @NotNull String getDisplayName() {
         return "HAnS";
     }
 }