+++ src/main/java/se/isselab/HAnS/syntaxHighlighting/codeAnnotations/CodeAnnotationsSyntaxHighlighter.java
@@ -55,9 +55,8 @@ public class CodeAnnotationsSyntaxHighlighter extends SyntaxHighlighterBase {
         return new CodeAnnotationLexerAdapter();
     }
 
-    @NotNull
     @Override
-    public TextAttributesKey[] getTokenHighlights(IElementType tokenType) {
+    public TextAttributesKey @NotNull [] getTokenHighlights(IElementType tokenType) {
         if (tokenType.equals(CodeAnnotationTypes.BEGIN) ||
                 tokenType.equals(CodeAnnotationTypes.END) ||
                 tokenType.equals(CodeAnnotationTypes.LINE))  {